Add decoration files and update the classes
