--- person.rb
@@ -1,21 +1,0 @@
-class Person
-  attr_accessor :name, :age
-  attr_reader :id
-
-  def initialize(name, age, parent_permission: true)
-    @id = Random.rand(1..1000)
-    @name = name
-    @age = age
-    @parent_permission = parent_permission
-  end
-
-  def can_use_services?
-    is_of_age? || @parent_permission
-  end
-
-  private
-
-  def of_age?
-    @age >= 18
-  end
-end
