--- student.rb
@@ -1,11 +1,0 @@
-require './person'
-class Student < Person
-  def initialize(name, age, parent_permission, classroom)
-    super(name, age, parent_permission, classroom)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    '¯\(ツ)/¯'
-  end
-end
